会员优惠

+++ pinet-rest/src/main/java/com/pinet/rest/entity/ScoreRecord.java
@@ -37,7 +37,7 @@ public class ScoreRecord extends BaseEntity {
     private String scoreTitle;
 
     @ApiModelProperty("积分")
-    private Integer score;
+    private Double score;
 
     @ApiModelProperty("用户等级 0普通用户 10普通会员  20店帮主")
     private Integer customerMember;
@@ -49,7 +49,7 @@ public class ScoreRecord extends BaseEntity {
     private Integer scoreType;
 
     @ApiModelProperty("当前积分")
-    private Integer customerScore;
+    private Double customerScore;
 
 
 }
